# backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Comprehensive Equipment & Place API",
    "version": "1.0.0",
    "description": "API documentation for a Flask application that includes Users (with FaceIdentity storing face image data), Equipment (and sub-entities like LedLCD, Light, Door), Control actions, and Place (location) endpoints."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Creates a NormalUser or AdminUser based on the provided role. To register as admin, include the correct admin_token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterUser" }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "Missing or invalid input" },
          "403": { "description": "Invalid admin token" }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "User login",
        "description": "Returns a JWT token upon successful authentication.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginUser" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": { "description": "Missing username or password" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/face_identity": {
      "post": {
        "summary": "Create or update FaceIdentity",
        "description": "Stores or updates the user's face identity with a face_id, name, and image data. Use multipart/form-data for file upload.",
        "security": [ { "BearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/FaceIdentityInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Face identity saved/updated" },
          "400": { "description": "Missing required fields" },
          "404": { "description": "User not found" }
        }
      },
      "get": {
        "summary": "Get FaceIdentity",
        "description": "Retrieves the user's face identity. The face image is returned as a Base64-encoded string.",
        "security": [ { "BearerAuth": [] } ],
        "responses": {
          "200": {
            "description": "FaceIdentity details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FaceIdentity" }
              }
            }
          },
          "404": { "description": "Face identity not found" }
        }
      }
    },
    "/places": {
      "post": {
        "summary": "Create a Place",
        "description": "Creates a new Place (room and address).",
        "security": [ { "BearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlaceInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Place created" },
          "400": { "description": "Invalid input" }
        }
      },
      "get": {
        "summary": "List all Places",
        "description": "Retrieves a list of all Place records.",
        "security": [ { "BearerAuth": [] } ],
        "responses": {
          "200": {
            "description": "Array of places",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Place" }
                }
              }
            }
          }
        }
      }
    },
    "/places/{place_id}": {
      "parameters": [
        {
          "name": "place_id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "description": "ID of the Place"
        }
      ],
      "get": {
        "summary": "Get a Place",
        "description": "Retrieves a Place along with its associated equipment.",
        "security": [ { "BearerAuth": [] } ],
        "responses": {
          "200": {
            "description": "Place details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PlaceWithEquipment" }
              }
            }
          },
          "404": { "description": "Place not found" }
        }
      },
      "put": {
        "summary": "Update a Place",
        "description": "Updates a Place's room and address.",
        "security": [ { "BearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlaceInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Place updated" },
          "404": { "description": "Place not found" }
        }
      },
      "delete": {
        "summary": "Delete a Place",
        "description": "Deletes the specified Place.",
        "security": [ { "BearerAuth": [] } ],
        "responses": {
          "200": { "description": "Place deleted" },
          "404": { "description": "Place not found" }
        }
      }
    },
    "/equipment": {
      "post": {
        "summary": "Create Equipment",
        "description": "Creates a new Equipment record, optionally linked to a Place.",
        "security": [ { "BearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EquipmentInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Equipment created" },
          "400": { "description": "Missing equipment name" }
        }
      },
      "get": {
        "summary": "List all Equipment",
        "description": "Retrieves all Equipment records.",
        "security": [ { "BearerAuth": [] } ],
        "responses": {
          "200": {
            "description": "Array of equipment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Equipment" }
                }
              }
            }
          }
        }
      }
    },
    "/equipment/{equipment_id}": {
      "parameters": [
        {
          "name": "equipment_id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "description": "ID of the Equipment"
        }
      ],
      "get": {
        "summary": "Get Equipment",
        "description": "Retrieves an Equipment record with its sub-devices.",
        "security": [ { "BearerAuth": [] } ],
        "responses": {
          "200": {
            "description": "Equipment details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EquipmentWithSubDevices" }
              }
            }
          },
          "404": { "description": "Equipment not found" }
        }
      },
      "put": {
        "summary": "Update Equipment",
        "description": "Updates an Equipment record.",
        "security": [ { "BearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EquipmentInput" }
            }
          }
        },
        "responses": {
          "200": { "description": "Equipment updated" },
          "404": { "description": "Equipment not found" }
        }
      },
      "delete": {
        "summary": "Delete Equipment",
        "description": "Deletes the specified Equipment record.",
        "security": [ { "BearerAuth": [] } ],
        "responses": {
          "200": { "description": "Equipment deleted" },
          "404": { "description": "Equipment not found" }
        }
      }
    },
    "/equipment/{equipment_id}/ledlcd": {
      "post": {
        "summary": "Create LedLCD",
        "description": "Creates a new LedLCD sub-device for the specified Equipment.",
        "security": [ { "BearerAuth": [] } ],
        "parameters": [
          {
            "name": "equipment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LedLCDInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "LedLCD created" },
          "400": { "description": "Missing name" },
          "404": { "description": "Equipment not found" }
        }
      },
      "get": {
        "summary": "List LedLCDs",
        "description": "Retrieves all LedLCD sub-devices for a given Equipment.",
        "security": [ { "BearerAuth": [] } ],
        "parameters": [
          {
            "name": "equipment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of LedLCD devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/LedLCD" }
                }
              }
            }
          },
          "404": { "description": "Equipment not found" }
        }
      }
    },
...
    "/controls": {
      "post": {
        "summary": "Create a Control action",
        "description": "Creates a new control action (e.g. open door, turn LED on) for the logged-in user.",
        "security": [ { "BearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ControlInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Control created" },
          "400": { "description": "Missing required fields" }
        }
      },
      "get": {
        "summary": "List Control actions",
        "description": "Retrieves all control actions for the logged-in user.",
        "security": [ { "BearerAuth": [] } ],
        "responses": {
          "200": {
            "description": "Array of control actions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Control" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterUser": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "example": "john_doe" },
          "password": { "type": "string", "example": "secret123" },
          "role": { "type": "string", "example": "user" },
          "name": { "type": "string", "example": "John Doe" },
          "account": { "type": "string", "example": "AC001" },
          "phone": { "type": "string", "example": "1234567890" },
          "action": { "type": "string", "example": "default_action" },
          "access": { "type": "string", "example": "full" },
          "admin_token": { "type": "string", "example": "mySuperSecretAdminToken123" }
        },
        "required": ["username", "password"]
      },
      "LoginUser": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "example": "john_doe" },
          "password": { "type": "string", "example": "secret123" }
        },
        "required": ["username", "password"]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..." }
        }
      },
      "FaceIdentityInput": {
        "type": "object",
        "properties": {
          "face_id": { "type": "string", "example": "FACE123" },
          "name": { "type": "string", "example": "John Doe" },
          "face_image": { "type": "string", "format": "byte", "example": "Base64EncodedImageData" }
        },
        "required": ["face_id"]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "BearerAuth": [] }
  ]
}